+++ changelog.md
@@ -10,6 +10,9 @@
 
 - Add `--skip-proto` flag to `build`, `init` and `serve` commands to build the chain without building proto files
 
+### Fixes
+
+- Fix `pkg/cosmosclient` to call the faucet prior to create the tx. (#2775)
 
 ## [`v0.23.0`](https://github.com/ignite/cli/releases/tag/v0.23.0)